+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/JdbcTomcatInstrumentedModule.java
@@ -19,11 +19,9 @@
 
 package io.bootique.jdbc.instrumented.tomcat;
 
-import javax.inject.Singleton;
-
 import com.codahale.metrics.MetricRegistry;
-import io.bootique.di.Binder;
 import io.bootique.di.BQModule;
+import io.bootique.di.Binder;
 import io.bootique.di.Provides;
 import io.bootique.jdbc.DataSourceFactory;
 import io.bootique.jdbc.JdbcModule;
@@ -31,6 +29,12 @@ import io.bootique.jdbc.instrumented.tomcat.healthcheck.TomcatHealthChecks;
 import io.bootique.metrics.MetricNaming;
 import io.bootique.metrics.health.HealthCheckModule;
 
+import javax.inject.Singleton;
+
+/**
+ * @deprecated The alternative is switching to bootique-jdbc-hikaricp.
+ */
+@Deprecated(since = "3.0", forRemoval = true)
 public class JdbcTomcatInstrumentedModule implements BQModule {
 
     public static final MetricNaming METRIC_NAMING = MetricNaming.forModule(JdbcTomcatInstrumentedModule.class);