+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #125 upgrade HikariCP to 5.0.1
 * #126 HikariCP - support for lazy DataSource creation
+* #127 Liquibase - conflict with Jakarta modules
 
 ## 3.0.M1