+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #125 upgrade HikariCP to 5.0.1
 * #126 HikariCP - support for lazy DataSource creation
 * #127 Liquibase - conflict with Jakarta modules
+* #129 Support Types.NUMERIC in DefaultFromStringConverter
 
 ## 3.0.M1